% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mono_bart.R
\name{fit_tree}
\alias{fit_tree}
\title{Fit A tree}
\usage{
fit_tree(df, minbucket)
}
\arguments{
\item{df}{Our integrals}

\item{minbucket}{How deep the tree is, larger is smaller depth of tree}
}
\description{
Fit A tree
}
\examples{

}
