% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mono_bart.R
\name{ggrocs}
\alias{ggrocs}
\title{Functions plots multiple 'roc' objects into one plot}
\usage{
ggrocs(rocs, breaks = seq(0, 1, 0.1), tittle = "Fit Model")
}
\arguments{
\item{rocs}{A list of 'roc' objects. Every list item has a name.}

\item{breaks}{A vector of integers representing ticks on the x- and y-axis}

\item{legentTitel}{A string which is used as legend titel}
}
\description{
Functions plots multiple 'roc' objects into one plot
}
